+++ lib/app/action_buttons/novel_home_action_button.dart
@@ -5,7 +5,7 @@ import 'package:novel_v3/app/dialogs/core/index.dart';
 import 'package:novel_v3/app/models/novel_model.dart';
 import 'package:novel_v3/app/riverpods/providers.dart';
 import 'package:novel_v3/app/route_helper.dart';
-import 'package:novel_v3/app/my_libs/novel_data/novel_data_scanner_screen.dart';
+import 'package:novel_v3/my_libs/novel_data/novel_data_scanner_screen.dart';
 
 class NovelHomeActionButton extends ConsumerStatefulWidget {
   const NovelHomeActionButton({super.key});