+++ lib/app/components/core/index.dart
@@ -1,5 +1,5 @@
 export './app_components.dart';
-export '../../my_libs/clean_cache/cache_component.dart';
+export '../../../my_libs/clean_cache/cache_component.dart';
 export './cover_components.dart';
 export './genres_list_view.dart';
 export './imdb_component.dart';