+++ lib/app/models/chapter_model.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:novel_v3/app/constants.dart';
-import 'package:novel_v3/app/my_libs/text_reader/text_reader_data_interface.dart';
+import 'package:novel_v3/my_libs/text_reader/text_reader_data_interface.dart';
 import 'package:than_pkg/than_pkg.dart';
 
 class ChapterModel extends TextReaderDataInterface {