+++ lib/app/models/index.dart
@@ -2,4 +2,4 @@ export 'core/app_config_model.dart';
 export './chapter_model.dart';
 export './novel_data_model.dart';
 export './novel_model.dart';
-export '../my_libs/pdf_readers/pdf_config_model.dart';
+export '../../my_libs/pdf_readers_v1.0.0/pdf_config_model.dart';