+++ lib/app/screens/app_setting_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:novel_v3/app/components/index.dart';
 import 'package:novel_v3/app/constants.dart';
 import 'package:novel_v3/app/dialogs/index.dart';
-import 'package:novel_v3/app/my_libs/general_server/proxy_hosting_server/index.dart';
+import 'package:novel_v3/my_libs/general_server/proxy_hosting_server/index.dart';
 import 'package:novel_v3/app/models/index.dart';
 import 'package:novel_v3/app/notifiers/app_notifier.dart';
 import 'package:novel_v3/app/services/index.dart';