+++ lib/app/services/core/index.dart
@@ -2,6 +2,6 @@ export 'android_app_services.dart';
 export 'app_config_services.dart';
 export 'app_path_services.dart';
 export 'recent_db_services.dart';
-export '../../my_libs/clean_cache/cache_services.dart';
+export '../../../my_libs/clean_cache/cache_services.dart';
 export 'dio_services.dart';
 export 'rabbit.dart';