+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:novel_v3/app/my_libs/general_server/index.dart';
+import 'package:novel_v3/my_libs/general_server/index.dart';
 import 'package:novel_v3/app/my_app.dart';
 import 'package:novel_v3/app/notifiers/app_notifier.dart';
 import 'package:novel_v3/app/services/index.dart';