+++ lib/my_libs/fetcher/types/fetcher.dart
@@ -1,5 +1,5 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:novel_v3/app/my_libs/fetcher/fetcher_query.dart';
+import 'package:novel_v3/my_libs/fetcher/fetcher_query.dart';
 
 class Fetcher {
   String title;