+++ lib/my_libs/pdf_readers_v1.0.0/pdfrx_reader_screen.dart
@@ -3,14 +3,14 @@ import 'dart:io';
 import 'package:cherry_toast/cherry_toast.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:novel_v3/app/my_libs/pdf_readers/pdf_bookmark_drawer.dart';
+import 'package:novel_v3/my_libs/pdf_readers_v1.0.0/pdf_bookmark_drawer.dart';
 import 'package:pdfrx/pdfrx.dart';
 import 'package:t_widgets/widgets/t_loader.dart';
 import 'package:than_pkg/enums/screen_orientation_types.dart';
 import 'package:than_pkg/than_pkg.dart';
 
-import '../../dialogs/index.dart';
-import '../../notifiers/app_notifier.dart';
+import '../../app/dialogs/index.dart';
+import '../../app/notifiers/app_notifier.dart';
 import 'pdf_config_model.dart';
 import 'pdf_reader_setting_dialog.dart';
 
@@ -63,17 +63,19 @@ class _PdfrxReaderScreenState extends State<PdfrxReaderScreen> {
   void onPdfLoaded() async {
     try {
       //set offset
-      final newOffset = Offset(oldOffsetX, oldOffsetY);
+
       if (oldZoom != 0 && oldOffsetX != 0 && oldOffsetY != 0) {
-        //delay
-        // await Future.delayed(const Duration(milliseconds: 800));
         await pdfController.goToPage(pageNumber: oldPage);
+        //delay
+
+        // await Future.delayed(const Duration(milliseconds: 500));
+
+        final newOffset = Offset(oldOffsetX, pdfController.centerPosition.dy);
+
         pdfController.setZoom(newOffset, oldZoom);
       } else {
         await pdfController.goToPage(pageNumber: oldPage);
       }
-      // final oldOffset = pdfController.centerPosition;
-      // final zoom = pdfController.currentZoom;
 
       await ThanPkg.platform
           .toggleFullScreen(isFullScreen: config.isFullscreen);